add data boostrap

+++ src/main/resources/db/migration/V1__CREATE_MENU_TABLE.sql
@@ -2,7 +2,7 @@ CREATE SCHEMA if not exists "restaurant";
 
 DROP TABLE IF EXISTS "menu";
 CREATE TABLE "menu"(
-    "id" BIGINT PRIMARY KEY,
+    "id" SERIAL PRIMARY KEY,
     "description" VARCHAR ,
     "price" VARCHAR(255) NOT NULL,
     "price2" VARCHAR (255),
